Markdown text processing interface and some package structure refactoring.
