+++ components-impl/src/main/java/ru/rougegibbons/landsanddungeons/components/core/numbers/longint/PairLongComponentImpl.java
@@ -3,7 +3,7 @@ package ru.rougegibbons.landsanddungeons.components.core.numbers.longint;
 import org.jetbrains.annotations.NotNull;
 import ru.rougegibbons.landsanddungeons.components.core.AbstractComponent;
 import ru.rougegibbons.landsanddungeons.components.core.numbers.generic.PairNumberComponentImpl;
-import ru.rougegibbons.landsanddungeons.components.interfaces.core.PairNumberComponent;
+import ru.rougegibbons.landsanddungeons.components.interfaces.core.numbers.PairNumberComponent;
 import ru.rougegibbons.landsanddungeons.components.models.ComponentModel;
 import ru.rougegibbons.landsanddungeons.components.models.core.PairLongComponentModel;
 import ru.rougegibbons.landsanddungeons.utils.constants.Constants;
