+++ components-impl/src/main/java/ru/rougegibbons/landsanddungeons/components/core/numbers/longint/SingleLongComponentImpl.java
@@ -3,7 +3,7 @@ package ru.rougegibbons.landsanddungeons.components.core.numbers.longint;
 import org.jetbrains.annotations.NotNull;
 import ru.rougegibbons.landsanddungeons.components.core.AbstractComponent;
 import ru.rougegibbons.landsanddungeons.components.core.numbers.generic.SingleNumberComponentImpl;
-import ru.rougegibbons.landsanddungeons.components.interfaces.core.SingleNumberComponent;
+import ru.rougegibbons.landsanddungeons.components.interfaces.core.numbers.SingleNumberComponent;
 import ru.rougegibbons.landsanddungeons.components.models.ComponentModel;
 import ru.rougegibbons.landsanddungeons.components.models.core.SingleLongComponentModel;
 import ru.rougegibbons.landsanddungeons.utils.constants.Constants;
