+++ components-impl/src/main/java/ru/rougegibbons/landsanddungeons/components/core/string/StringComponentImpl.java
@@ -2,7 +2,7 @@ package ru.rougegibbons.landsanddungeons.components.core.string;
 
 import org.jetbrains.annotations.NotNull;
 import ru.rougegibbons.landsanddungeons.components.core.AbstractComponent;
-import ru.rougegibbons.landsanddungeons.components.interfaces.core.StringComponent;
+import ru.rougegibbons.landsanddungeons.components.interfaces.core.string.StringComponent;
 import ru.rougegibbons.landsanddungeons.components.models.ComponentModel;
 import ru.rougegibbons.landsanddungeons.components.models.core.StringComponentModel;
 import ru.rougegibbons.landsanddungeons.utils.constants.Constants;
